requirements.txt
addict==2.4.0
appdirs==1.4.4
asynctest==0.13.0
attributee==0.1.5
attrs==21.4.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
cityscapesScripts==2.2.0
click==7.1.2
codecov==2.1.12
colorama==0.4.4
coloredlogs==15.0.1
coverage==6.4.1
cycler==0.11.0
Cython==0.29.30
dotty-dict==1.3.0
flake8==4.0.1
flatbuffers==2.0
fonttools==4.33.3
humanfriendly==10.0
idna==3.3
imagecorruptions==1.1.2
imageio==2.19.3
importlib-metadata==4.11.4
iniconfig==1.1.1
interrogate==1.5.0
isort==4.3.21
joblib==1.1.0
kiwisolver==1.4.2
kwarray==0.6.0
lap==0.4.0
Markdown==3.3.7
matplotlib==3.5.2
mccabe==0.6.1
mmcls==0.23.1
mmcv-full==1.5.2
mmdet==2.25.0
mmtrack==0.13.0
model-index==0.1.11
motmetrics==1.2.5
networkx==2.8.3
numpy==1.22.4
onnx==1.7.0
onnxruntime==1.11.1
opencv-python==4.5.5.64
openmim==0.1.5
ordered-set==4.1.0
packaging==21.3
pandas==1.3.5
Pillow==9.1.1
pip==21.2.4
pluggy==1.0.0
protobuf==3.20.1
py==1.11.0
pycocotools==2.0.2
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.9
pyquaternion==0.9.9
pytest==7.1.2
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
requests==2.27.1
scikit-image==0.19.2
scikit-learn==1.1.1
scipy==1.7.3
seaborn==0.11.2
setuptools==61.2.0
setuptools-scm==6.4.2
six==1.16.0
sklearn==0.0
tabulate==0.8.9
terminaltables==3.1.10
threadpoolctl==3.1.0
tifffile==2022.5.4
timm==0.5.4
toml==0.10.2
tomli==2.0.1
torch==1.10.0+cu111
torchaudio==0.10.0+rocm4.1
torchvision==0.11.0+cu111
tqdm==4.64.0
typing==3.7.4.3
typing_extensions==4.2.0
ubelt==1.1.0
urllib3==1.26.9
wheel==0.37.1
xdoctest==1.0.0
xmltodict==0.13.0
yapf==0.32.0
zipp==3.8.0
